+++ modules/juce_audio_basics/midi/juce_MidiBuffer.cpp
@@ -34,7 +34,7 @@ namespace MidiBufferHelpers
 
     inline uint16 getEventTotalSize (const void* const d) noexcept
     {
-        return getEventDataSize (d) + sizeof (int32) + sizeof (uint16);
+        return (uint16) (getEventDataSize (d) + sizeof (int32) + sizeof (uint16));
     }
 
     static int findActualEventLength (const uint8* const data, const int maxBytes) noexcept
